requirements.txt
fdb
gspread
oauth2client
schedule
python-dotenv
fdb
gspread-formatting
pyinstaller
